Couple quick fixes to UniConf--removed non-existent constructor from UniConfClient header file and fixe UniConfClient destructor.  Well, dgtaylor did, actually.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@1538 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ uniconf/uniconfclient.cc
@@ -30,7 +30,8 @@ UniConfClient::~UniConfClient()
     {
         if (conn->isok())
             conn->write(WvString("%s\n", UniConfConn::UNICONF_QUIT));
-        list->unlink(conn);
+	if (list)
+	    list->unlink(conn);
         delete conn;
      }
 }
